+++ backend/src/main/java/com/fung/fungry/ModelDTO/CartDTO.java
@@ -14,6 +14,6 @@ public class CartDTO {
 
     private Long cartId;
     private List<CartItemDTO> cartItemDTOS;
-    private Double totalAmt;
+    private Long totalAmt;
     private String restaurantName;
 }
