+++ backend/src/main/java/com/fung/fungry/ModelDTO/OrderDTO.java
@@ -21,7 +21,7 @@ public class OrderDTO {
     private Integer expecetedTimeInMinutes;
     private String restaurantName;
     private OrderStatus status;
-    private Double totalAmt;
+    private Long totalAmt;
     private LocalDateTime createdTime;
 
 }
